requirements.txt
Mezzanine>=3.1.10
cartridge>=0.9.5
django-pinpayments
